Change GeoIP field name for dest_ip to not overlap with GeoIP for src_ip
This commit renames the `geoip` field for `dest_ip` to `geoip_dest` in
`logstash.conf` and creates a new mapping in `elasticsearch6-template.json`.

In logstash configuration, GeoIP is used for both `dest_ip` and `src_ip`
if both are present. However, the `target` field have the same name
(`geoip`) for both, so in the end the `dest_ip` GeoIP will overwrite the
`src_ip` one.